moved globals
