+++ cellprofiler/gui/view_workspace/control/_object_list.py
@@ -1,7 +1,7 @@
 import cellprofiler_core.preferences
 
 import cellprofiler.gui
-from cellprofiler.gui.view_workspace import bind_data_class
+from cellprofiler.gui.view_workspace.control import bind_data_class
 from cellprofiler.gui.view_workspace.control._control import Control
